Inject telemetry router grain via DI
